fix(cloud): carry the size refusal in the classifier, not in two pre-emptions
The oversized-restore fix already failed closed, but it did so by placing an
instanceof check in front of each of the two call sites that consult
isUnrecoverableSnapshotError. The classifier itself still answered true, so a
third call site added later would silently re-enter the destroy-and-boot-fresh
degrade for a chain that is intact and decryptable. The classifier now answers
false and the refusal keeps its own terminal branch, so the invariant travels
with the error rather than with the order of two if-statements.

Two further gaps in the same fix:

The fetch-side branch was unreachable in production. Backup-chain
reconstruction is the real oversize refusal on that path and it threw a plain
Error, so a genuine chain-budget breach exited as an untyped rethrow with no
consent guidance, while the test that covered the branch injected a typed error
production never produced. The repository now throws the typed class. It lives
in the shared limits module beside the limits themselves -- both sides that
enforce them throw it, and a service-owned class would make the repository
import its own consumer.

The size branch was ordered after the from-backup rethrow, and the gated wake
always passes from-backup. So on a wake -- the one path where the forceFreshBoot
consent mechanism exists -- the operator got the bare typed message instead of
the sentence naming the consent. Reordered at both sites.

Tests follow the classification rather than pin the old answer: the refusal is
now asserted to be neither unrecoverable nor permanently lost, since the first
authorises the fresh-boot degrade and the second additionally authorises
pruning the chain.
